+++ mobile/android/android-components/docs/changelog.md
@@ -23,6 +23,7 @@ permalink: /changelog/
 
 * **feature-prompts**:
   * 🚒 Bug fixed [issue #9229](https://github.com/mozilla-mobile/android-components/issues/9229) - Dismiss SelectLoginPrompt from the current tab when opening a new one ensuring the new one can show it's own. When returning to the previous tab users should focus a login field to see the SelectLoginPrompt again.
+  * PromptFeature now implements UserInteractionHandler.onBackPressed to dismiss loginPicker.
 
 * **feature-downloads**:
   * Allow browsers to change the download notification accent color by providing `Style()` in `AbstractFetchDownloadService`, for more information see [#9299](https://github.com/mozilla-mobile/android-components/issues/9299).